+++ labpack/email/__init__.py
@@ -1,0 +1,3 @@
+__author__ = 'rcj1492'
+__created__ = '2017.03'
+__license__ = 'MIT'