+++ labpack/records/id.py
@@ -83,7 +83,7 @@ class labID(object):
         self.mac = local_mac
 
 if __name__ == '__main__':
-    print(labID().id48)
+    print(labID().id24)
     print(labID().epoch)